blame-tree: add config to add to revopts
This is a terrific hack, and should be reverted when we come
up with a better solution.

Basically, some repositories are so big that the full
blame-tree run takes longer than our reading-side timeout
allows.  Blame-tree was designed to produce incremental
results, but the reading side simply throws away the
incremental results when we hit the timeout.

To solve this, we're going to limit traversals manually in a
few repositories. We can do this easily by adding
"--since=5.years.ago" or similar to the command-line, but
the simplest way to do that is to stick it in git-config.
And so we have this patch.