config.mak.local: modify GIT_TEST_OPTS
Setting '-x' and '--verbose-log' causes a defect in the 't/perf' tests
which causes all results to report as '<missing>'. This is known (and
fixed) upstream in 0baf78e7bc (perf-lib.sh: rely on test-lib.sh for
--tee handling, 2019-03-15).

For now--or at least when working on forks of git.git that don't have
the above--set an override of 'GIT_TEST_OPTS' which does not contain
these options, and thus allows the 't/perf' to run correctly.