contrib/git-jump/git-jump: support alias expansion
When a caller of 'git-jump' has, say, the alias 'di' assigned to 'diff',
it can be cumbersome to remember to type 'git jump diff' instead of the
shorthand '... di' that they are used to.

Let's teach 'git-jump' to expand these aliases before calling the
mode-specific subroutine. Do so by fetching the configuration value of
'alias.$1', defaulting to "$1" in the case that no alias is set.

Signed-off-by: Taylor Blau <me@ttaylorr.com>