commit-graph.c: avoid unnecessary tag dereference when merging
When performing a 'git commit-graph write' with '--split', the
commit-graph machinery calls 'merge_commit_graph()' after deciding on a
split strategy to optionally clean up any existing commit-graph
layers that were made obsolete by the split strategy [1].

At this time, 'merge_commit_graph()' checks each commit that it writes
into the merged graph to make sure that it still exists in the object
store.

To do this, it uses 'lookup_commit_reference_gently()', which accepts
either a commit object, or a tag that refers to a commit. However, since
all 'oid' arguments passed to this function are from within the
commit-graphs being merged, we never pass a commit reference, and so any
time we spend in 'deref_tag()' is wasted.

Improve the situation by using 'repo_has_object_file_with_flags' to
check if the object still exists, and '{lookup,repo_parse}_commit()' to
turn it into a bona-fide 'struct commit *'.

While we're at it, pass the 'SKIP_FETCH_OBJECT' flag to
'repo_has_object_file_with_flags' in order to avoid fetching non-local
commits. Such commits can go missing in a partially cloned repository
that has undergone a 'git gc'. In this case, avoid an unnecessary fetch
by simply dropping those commits from the merged graph.

In my testing environment, this improves the time to "merge" a split
commit-graph significantly. To test, we begin with a commit-graph chain
containing all reachable commits in the kernel. Then, we create one more
commit, and generate another incremental commit-graph on top, and force
it to be merged down with '--split=merge-all'.

To reproduce the following numbers, follow these steps:

  $ graphdir=".git/objects/info/commit-graphs"
  $ c="$(git rev-parse HEAD)"
  $ best-of-five -p "
      rm -rf $graphdir && (
        git reset --hard $c &&
        git commit-graph write --split=no-merge --reachable &&
        git commit --allow-empty -mfoo
      ) >/dev/null 2>&1"
      sh -c "
        git rev-list HEAD^..HEAD |
        git commit-graph write --split=merge-all --stdin-commits"

On my workstation, this improves the timings from:

  Attempt 1: 18.188
  Attempt 2: 18.047
  Attempt 3: 18.139
  Attempt 4: 18.016
  Attempt 5: 18.544

  real	0m18.016s
  user	0m17.480s
  sys	0m0.500s

(where the above is based on the current tip of 'next') to:

  Attempt 1: 5.458
  Attempt 2: 5.432
  Attempt 3: 4.711
  Attempt 4: 5.468
  Attempt 5: 5.081

  real	0m4.711s
  user	0m4.476s
  sys	0m0.212s

(after applying this patch) for a best-case savings of ~73.9%, and an
average savings of ~71.2%.

[1]: This can happen if, for example, the new commit-graph exceeds the
     maximum allowed factor on the number of commits.

Co-authored-by: Jeff King <peff@peff.net>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Taylor Blau <me@ttaylorr.com>