t/t5324: support extra chunks in 'graph_read_expect'
The 'graph_read_expect' function wraps the 'graph-read' test helper in
order to assert various properties of commit-graphs generated during
test, such as the number of commits in the graph, what chunks it
contains, and so on.

In a subsequent commit, we will add a test for which the generated graph
has an 'extra_edges' chunk. To support this, allow for an optional third
argument which specifies any non-standard chunks, such as extra-edges.