pack-bitmap.c: generify filtering functions
In 4f3bd5606a (pack-bitmap: implement BLOB_NONE filtering, 2020-02-14),
filtering support for bitmaps was added for the 'LOFC_BLOB_NONE' filter.

In the future, we would like to add support for filters that behave as
if they exclude a certain type of object, for e.g., the tree depth
filter with depth 0.

To prepare for this, generify some of the functions used for filtering,
such as 'find_tip_blobs' and 'filter_bitmap_blob_none' to work over
arbitrary object types.

To that end, create 'find_tip_objects' and
'filter_bitmap_exclude_type', and redefined the aforementioned functions
in terms of those.

Signed-off-by: Taylor Blau <me@ttaylorr.com>