commit-graph.c: introduce '--[no-]verify-oids'
When operating on a stream of commit OIDs on stdin, 'git commit-graph
write' checks that each OID refers to an object that is indeed a commit.
This is convenient to make sure that the given input is well-formed, but
can sometimes be undesirable.

For example, server operators may wish to feed the full commit object
IDs pointed to by refs that were updated during a push to 'git
commit-graph write --input=stdin-commits', and silently discard any
input that doesn't point at a commit. This can be done by combing the
output of 'git for-each-ref' with '--format %(*objecttype)', but this
requires opening up a potentially large number of objects. Instead, it
is more convenient to feed the updated object IDs to the commit-graph
machinery, and let it throw out whatever remains.

Introduce '--[no-]verify-oids' to make such a behavior possible. With
'--verify-oids' (the default behavior to retain backwards compatibility),
'git commit-graph write' will barf on a non-commit line in its input.
With '--no-verify-oids', such lines will be silently ignored, making the
above possible by specifying this option.

No matter which is supplied, 'git commit-graph write' retains the
behavior from the previous commit of rejecting non-OID inputs like
"HEAD" and "refs/heads/foo" as before. Well-formed OIDs that do not
refer to any existing object are errors in either case, since it is
likely a caller in this situation raced with some other object deletion,
and they may want to be alerted as such.

Signed-off-by: Taylor Blau <me@ttaylorr.com>