commit-graph.c: introduce '--[no-]check-oids'
When operating on a stream of commit OIDs on stdin, 'git commit-graph
write' checks that each OID refers to an object that is indeed a commit.
This is convenient to make sure that the given input is well-formed, but
can sometimes be undesirable.

For example, server operators may wish to feed the refnames that were
updated during a push to 'git commit-graph write --input=stdin-commits',
and silently discard refs that don't point at commits. This can be done
by combing the output of 'git for-each-ref' with '--'%(*objecttype)',
but this requires opening up a potentially large number of objects.
Instead, it is more convenient to feed the updated refs to the
commit-graph machinery, and let it throw out refs that don't point to
commits.

Introduce '--[no-]check-oids' to make such a behavior possible. With
'--check-oids' (the default beahvior to retain backwards compatability),
'git commit-graph write' will barf on a non-commit line in its input.
With 'no-check-oids', such lines will be silently ignored, making the
above possible by specifying this option.

Signed-off-by: Taylor Blau <me@ttaylorr.com>