commit-graph.c: don't check OIDs with '--stdin-commits'
In 7c5c9b9c57 (commit-graph: error out on invalid commit oids in 'write
--stdin-commits', 2019-08-05), the commit-graph machinery learned to
error out with '--stdin-commits' on OIDs that referred to non-commits.

7c5c9b9c57 makes a point of distinguishing between input from
'--stdin-commits' and '--reachable'. In the '--reachable' case,
non-commits are ignored, since the commit-graph machinery considers all
refs, including tags that point to non-commits.

This can be an inconvenience if you are using '--stdin-commits' like
'--reachable'. This can occur when a caller only passes references that
they know have been updated (for e.g., after a push). To skirt around
this, the caller must pay the price of opening up every ref (either via
'--reachable', or by checking the '%(*objecttype)' in the output of 'git
for-each-ref').

This is undesirable behavior for callers who are operating in
repositories that have a potentially very large number of references,
and who rely upon passing input to '--stdin-commits' to limit the number
of object inspections that are done.

To make this situation more convenient, drop the
'COMMIT_GRAPH_WRITE_CHECK_OIDS' flag, instead having the commit-graph
machinery silently discard non-commit objects, as was the case prior to
7c5c9b9c57.

Suggested-by: Derrick Stolee <dstolee@microsoft.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>