commit-graph.c: lookup commits outside of 'fill_oids_from_commits'
The commit-graph machinery uses 'fill_oids_from_commits' to peel the
list of commits and/or tags down into a set of OIDs. It is this set that
the commit-graph machinery will later take the reachability closure over
in order to build the final graph.

This patch makes it the caller's responsibility to peel the input before
passing it to 'fill_oids_from_commits'. There are two cases here to deal
with:

  * When the input comes from '--stdin-commits', the object is parsed
    in the builtin, and an error is thrown if the object does not exist.
    (This makes it much easier to determine if the object actually
    exists or not, because we can't reliably conclude the same from the
    old 'lookup_commit_reference_gently' returning 'NULL', since it
    might have come from any referenced object).

    This code is more or less the same as it was in
    'fill_oids_from_commits'. Since the object is parsed during the
    builtin, 'lookup_commit' becomes a noop within
    'fill_oids_from_commits'.

    Note that this is the case that can potentially take a long time,
    since it's proportional to the size of the input. For this reason,
    the progress meter is preserved in this case.

  * When the input comes from '--reachable', we are building up input in
    callbacks from 'for_each_ref', and can use 'peel_ref' immediately
    and avoid dereferencing a layer of tags, instead relying on the
    packed-refs file, if it exists.

    This case is quick, and therefore doesn't need a progress meter.

Splitting 'fill_oids_from_commits' into these two cases allows us to use
and not use a progress meter where it makes sense to do so, and likewise
makes the implementation of 'fill_oids_from_commits' much simpler.

Suggested-by: Jeff King <peff@peff.net>
Signed-off-by: Taylor Blau <me@ttaylorr.com>