blame-tree: replace string_list with a hashmap
Our implementation of 'git blame-tree' keeps track of a list of
interesting paths by storing them in lexicographic order in a
string_list.

Since our time is most often spent in computing diffs, the 'O(log n)'
lookup time gets dwarfed in all but exceptional cases. In a subsequent
commit, we will compute far fewers diffs on average by making use of the
changed-path Bloom filters when available.

This change has the potential to make that 'O(log n)' term much more
noticable. In preparation for that, make lookups constant time by
replacing the string_list with a hashmap. In the interim, this doesn't
change performance much (it is perhaps slightly better, but really the
results are dominated by time spent computing diffs).

Before this patch, running 'git blame-tree .' in the kernel (this and
the time below it are best-of-five):

    real	1m14.832s
    user	1m14.450s
    sys	0m0.236s

after:

    real	1m14.056s
    user	1m13.628s
    sys	0m0.272s

Co-authored-by: Jeff King <peff@peff.net>