revision.c: avoid recomputing strlen
In 'prepare_to_use_bloom_filter', we sometimes munge the pathspec, and
other times do not. In the cases where we do not, the length of the spec
is already computed, and so we do not need to waste time computing it
again.

Divide these cases and reuse the already-computed value when it is
available.

Signed-off-by: Taylor Blau <me@ttaylorr.com>