revision.c: avoid querying empty Bloom keys
When preparing to use Bloom filters during a revision walk with a
pathspec, we normalize the given spec before generating and querying the
corresponding Bloom key.

This works incorrectly when the given pathspec is '.', and the revision
walk was started from the root. Normalizing '.' with the repository root
yields the empty string, which is never inserted as a Bloom key into the
filter.

In this case, the correct thing to do is to neither fill nor query a
Bloom key at all, in which case we check that the resulting pathspec is
non-empty before forming a Bloom key based on it.

Helped-by: Derrick Stolee <dstolee@microsoft.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>