commit.c: don't persist substituted parents when unshallowing
In 37b9dcabfc (shallow.c: use '{commit,rollback}_shallow_file',
2020-04-22), Git learned how to reset stat-validity checks for the
'$GIT_DIR/shallow' file, allowing it to change between a shallow and
non-shallow state in the same process (e.g., in the case of 'git fetch
--unshallow').

However, 37b9dcabfc does not alter or remove any grafts nor substituted
parents. This produces a problem when unshallowing if another part of
the code relies on the un-grafted and/or non-substituted parentage for
commits after, say, fetching.

This can arise 'fetch.writeCommitGraph' is true. Ordinarily (and
certainly previous to 37b9dcabfc), 'commit_graph_compatible()' would
return 0, indicating that the repository should not generate
commit-graphs (since at one point in the same process it was shallow).
But with 37b9dcabfc, that check succeeds.

This is where the bug occurs: even though the repository isn't shallow
any longer (that is, we have all of the objects), the in-core
representation of those objects still has munged parents at the shallow
boundaries. If a commit-graph write proceeds, we will use the incorrect
parentage, producing wrong results.

(Prior to this patch, there were two ways of fixing this: either (1)
set 'fetch.writeCommitGraph' to 'false', or (2) drop the commit-graph
after unshallowing).

One way to fix this would be to reset the parsed object pool entirely
(flushing the cache and thus preventing subsequent reads from modifying
their parents) after unshallowing. This can produce a problem when
callers have a now-stale reference to the old pool, and so this patch
implements a different approach. Instead, we attach a new bit to the
pool, 'substituted_parent' which indicates if the repository *ever*
stored a commit which had its parents modified (i.e., the shallow
boundary *before* unshallowing).

This bit is sticky, since all subsequent reads after modifying a
commit's parent are unreliable when unshallowing. This patch modifies
the check in 'commit_graph_compatible' to take this bit into account,
and correctly avoid generating commit-graphs in this case.

Helped-by: Derrick Stolee <dstolee@microsoft.com>
Helped-by: Jonathan Nieder <jrnieder@gmail.com>
Reported-by: Jay Conrod <jayconrod@google.com>
Signed-off-by: Taylor Blau <me@ttaylorr.com>