blame-tree: limit traversal by Bloom keys too
In 'git blame-tree --go-faster', we use a home-grown revision walk in
order to implement our walk, which is non-standard and not possible with
the existing revision walking API.

When we visit a commit, we call 'maybe_changed_path()' before computing
a diff, to see if that commit is relevant to our results or not. Prior
to this patch, we checked whether any Bloom key corresponding to our set
of active paths was contained in the current commit's Bloom filter.

Critically, we did *not* check whether the commit changed *any* paths in
our pathspec before checking whether or not the set of active paths were
changed.

Introduce 'blame_tree_fast_init_bloom()' to populate those keys by
stealing the logic from 'revision.c:prepare_to_use_bloom_filter()'.
Actually usuing that function isn't so straightforward, because of how
we munge variables in the rev_info struct, so some copying occurs here.
(If we did decide to upstream these patches, that would be worth getting
rid of).

Once we populate those keys, check in 'maybe_changed_path()' whether our
pathspec is modified *first*, before checking whether or not the
still-active paths were modified, too.

This provides an improvement on the status-quo for '--go-faster' with
Bloom filters by a substantial margin. At the previous patch (in
DefinitelyTyped/DefinitelyTyped's 'types/abbrev' directory), our numbers
look something like:

  $ git --version
  git version 2.27.0.2960.g562003b90d
  $ best-of-five git blame-tree --max-depth=2 types/abbrev
  real	0m0.183s
  user	0m0.148s
  sys	0m0.036s
  $ best-of-five git blame-tree --go-faster --max-depth=2 types/abbrev
  real	0m0.484s
  user	0m0.412s
  sys	0m0.072s

Compared with this patch, which drops the latter timings down by ~43.7%:

  $ git --version
  git version 2.27.0.2960.gb33187b440
  $ best-of-five git blame-tree --go-faster --max-depth=2 types/abbrev
  real	0m0.086s
  user	0m0.053s
  sys	0m0.033s

I'm hoping that this resolves our long-standing conjecture that there is
"more overhead per-commit, but fewer commits overall" in the direction
of "the overhead is nominal, and we had a bug in our revision walk which
has since been resolved".

Co-authored-by: Derrick Stolee <dstolee@microsoft.com>