blame-tree: hackily detect non-commits in revision walk
A recent commit switched to using prepare_revision_walk(), which
discards any non-commits, and we lost our ability to complain about
being passed a tree, etc. But we can ask for those non-commits to be
left in the pending array by specifying rev_info.blob_objects, etc. This
is a bit of a hack, since we know we would not want to actually traverse
those objects, but:

  - we'll die() immediately if prepare_revision_walk() puts anything
    there

  - we're not using the traversal machinery in list-objects.c; our
    custom traversal ignores them completely

Note that we _don't_ ask for tag objects. Those are acceptable if they
point to a commit, so complaining about them in pending would be wrong.
And a tag-to-tree or tag-to-blob will still trigger our check because it
will leave the pointed-to object in the pending array.