blame-tree: complain about unknown arguments
When we parse arguments to 'git blame-tree', we call 'setup_revisions()'
which leaves any arguments that it doesn't recognize in argv. But, we
don't recognize those arguments either (since 'git blame-tree' doesn't
take any non-standard arguments).

We used to rely on not complaining about unknown arguments back when we
had '--go-faster'. Now that that is gone, we no longer rely on this
quirk, and can start complaining about unknown arguments.

We could get fancier about how we report these, but picking off the
first unknown argument should be sufficient.