midx: expose 'write_midx_file_only()' publicly
Expose a variant of the write_midx_file() function which ignores packs
that aren't included in an explicit "allow" list.

This is useful for callers which want to write a MIDX containing _some_
of the packs in the object store but not all of them. For example,
immediately after repacking, but before the redundant packs have been
removed.