builtin/repack.c: support writing a MIDX while repacking
Teach 'git repack' a new '--write-midx' option for callers that wish to
persist a multi-pack index in their repository while repacking.

There are two existing alternatives to this new flag, but they don't
cover our particular use-case. These alternatives are:

  - Call 'git multi-pack-index write' after running 'git repack', or

  - Set 'GIT_TEST_MULTI_PACK_INDEX=1' in your environment when running
    'git repack'.

The former introduces a window of time between 'git repack' and 'git
multi-pack-index write' where the repository is without a bitmap.
Consider the following:

  1. The caller runs 'git repack -Ad', creating a new packfile of all
     objects, and dropping all other packfiles.

  2. Then the caller runs 'git multi-pack-index write --bitmap', which
     creates a MIDX, and writes a bitmap over that MIDX.

This is insufficient because there is a window of time between (1) and
(2) where no bitmap exists at all. On the other hand, the invocation in
(1) could be modified to instead say 'git repack -Adb', in which case
there would always be a bitmap, but it would be a waste of CPU time
since we're generating effectively the same bitmap twice.

The latter (setting the 'GIT_TEST_' environment) variable is obviously
unsupported and not a good idea. In fact, even if it were supported
officially, it still wouldn't work, because it generates the MIDX
*after* redundant packs have been dropped, leading to the same issue
as above.

Introduce a new option which eliminates this race by teaching 'git
repack' to generate the MIDX at the critical point: after the new packs
have been written and moved into place, but before the redundant packs
have been removed.

This option is compatible with 'git repack''s '--bitmap' option (it
changes the interpretation to be: "write a bitmap corresponding to the
MIDX after one has been generated").

There is a little bit of additional noise in the patch below to avoid
repeating ourselves when selecting which packs to delete. Instead of a
single loop as before (where we iterate over 'existing_packs', decide if
a pack is worth deleting, and if so, delete it), we have two loops (the
first where we decide which ones are worth deleting, and the second
where we actually do the deleting). This makes it so we have a single
check we can make consistently when (1) telling the MIDX which packs we
want to exclude, and (2) actually unlinking the redundant packs.